[server] Moved MonitoredMetricWrapper to ai.startree.thirdeye.core inside thirdeye-server
