+++ thirdeye-server/src/main/java/ai/startree/thirdeye/core/MonitoredMetricWrapper.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package ai.startree.thirdeye.alert;
+package ai.startree.thirdeye.core;
 
 import com.google.common.base.Objects;
